backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
motor==3.6.0
pymongo==4.10.0
bcrypt==4.2.0
PyJWT==2.9.0
python-dotenv==1.0.1
python-multipart==0.0.12
